Added task to destroy all assets in database.
When a file already exists in the database and an import is run, the file ends
up in the database twice.  Using destroy will delete all files from the
database and then the imported assets will not be duplicated.

This is a sledgehammer approach, but it works, for example, when trying to dump
a production site to develop locally.